Update to 10214 - Plus WARNING
WARNING: this model is larger than most to the point that two gigabytes of Graphics memory can barely handle the initial brick fall... don't ask me why motion should be the the straw that breaks the memory's back but that seems to be the limiting factor.
ALSO - there is a minor graphics glitch with at least three bricks in this build; the curved tower base bricks, the splayed tower base bricks and the road surface flats all do not have icons in the inventory, just the plain colour that the brick is. I have yet to determine if this is a memory issue or because these bricks themselves have issues.

Updated 10214 - complete but for shield stickers
- fixed mistaken dependency
- adjusted containers due to size and distances when building the model from the ground