Moved stickered bricks to the bottom, updated counts and added various bricks from sets not yet added
Moved stickered bricks to the bottom, updated counts and added various bricks from sets not yet added

(NB: When the number 1 is postfixed with a "." this is where 1 instance of this brick missing is confirmed and the number of known instances is exactly one. If the number is greater than one no "." is postfixed but that number of instances is confirmed to be needed in this git repo... this is because in the old not counted list there had to be at least one instance and now that I'm counting them I'm confirming that one before adding one to the count to avoid an off by one! *sigh* - should have counted them into this list in the first place!)